+++ nc2zarr/writer.py
@@ -22,6 +22,7 @@
 import datetime
 import json
 import os.path
+from enum import Enum
 from typing import Dict, Any, Sequence
 
 import fsspec
@@ -31,13 +32,24 @@ import pandas as pd
 import retry.api
 import xarray as xr
 import zarr
+from zarr.errors import GroupNotFoundError
 
 from .constants import DEFAULT_OUTPUT_APPEND_DIM_NAME
 from .constants import DEFAULT_OUTPUT_RETRY_KWARGS
 from .custom import load_custom_func
+from .dataslice import append_slice
 from .log import LOGGER
 from .log import log_duration
+from .dataslice import find_slice
+from .dataslice import update_slice
 from .version import version
+
+
+_APPEND_MODES = ["all", "no_overlap", "newer", "replace", "retain"]
+AppendMode = Enum("AppendMode", zip(_APPEND_MODES, _APPEND_MODES))
+
+# TODO: Refactor this class, because it has become confusing.
+#   (See https://github.com/bcdev/nc2zarr/pull/44 )
 
 
 class DatasetWriter:
@@ -50,6 +62,7 @@ class DatasetWriter:
                  output_overwrite: bool = False,
                  output_append: bool = False,
                  output_append_dim: str = None,
+                 output_append_mode: AppendMode = AppendMode.all,
                  output_adjust_metadata: bool = False,
                  output_metadata: Dict[str, Any] = None,
                  output_s3_kwargs: Dict[str, Any] = None,
@@ -62,19 +75,24 @@ class DatasetWriter:
             raise ValueError('output_path must be given')
         if output_append and output_custom_postprocessor:
             raise ValueError('output_append and output_custom_postprocessor'
-                             ' cannot be given both')
+                             ' cannot both be given')
+
         self._output_path = output_path
-        self._output_custom_postprocessor = load_custom_func(output_custom_postprocessor) \
+        self._output_custom_postprocessor = \
+            load_custom_func(output_custom_postprocessor) \
             if output_custom_postprocessor else None
         self._output_consolidated = output_consolidated
         self._output_encoding = output_encoding
         self._output_overwrite = output_overwrite
         self._output_append = output_append
-        self._output_append_dim = output_append_dim or DEFAULT_OUTPUT_APPEND_DIM_NAME
+        self._output_append_dim =\
+            output_append_dim or DEFAULT_OUTPUT_APPEND_DIM_NAME
         self._output_adjust_metadata = output_adjust_metadata
+        self._output_append_mode = output_append_mode
         self._output_metadata = output_metadata
         self._output_s3_kwargs = output_s3_kwargs
-        self._output_retry_kwargs = output_retry_kwargs or DEFAULT_OUTPUT_RETRY_KWARGS
+        self._output_retry_kwargs =\
+            output_retry_kwargs or DEFAULT_OUTPUT_RETRY_KWARGS
         self._input_decode_cf = input_decode_cf
         self._input_paths = input_paths
         self._finalize_only = finalize_only
@@ -145,6 +163,9 @@ class DatasetWriter:
 
     def _append_dataset(self, ds: xr.Dataset):
         with log_duration('Appending dataset'):
+
+            self._check_append_allowed(ds)
+
             # Fix for https://github.com/bcdev/nc2zarr/issues/38
             # Get rid of variables that lack append_dim dimension:
             append_dim = self._output_append_dim
@@ -173,11 +194,94 @@ class DatasetWriter:
                 ds = self._remove_variable_attrs(ds)
 
             if not self._dry_run:
-                ds.to_zarr(self._output_store,
-                           append_dim=append_dim,
-                           consolidated=self._output_consolidated)
+                mode = self._output_append_mode
+                # From Python 3.10 on, the following would be better implemented
+                # with structural pattern matching (PEPs 634 - 636).
+                if mode in (AppendMode.replace, AppendMode.retain):
+                    self._append_with_insertions(ds)
+                elif mode is AppendMode.newer:
+                    output_ds = xr.open_zarr(self._output_store)
+                    # NB: assumes both ds and output_ds increasing in append_dim
+                    ds_new_only = ds.where(
+                        ds[append_dim] > output_ds[append_dim][-1],
+                        drop=True)
+                    ds_new_only.to_zarr(self._output_store,
+                                        append_dim=append_dim,
+                                        consolidated=self._output_consolidated)
+                elif mode in (AppendMode.all, AppendMode.no_overlap):
+                    ds.to_zarr(self._output_store,
+                               append_dim=append_dim,
+                               consolidated=self._output_consolidated)
+                else:
+                    # This should never happen, but best to handle it anyway.
+                    raise NotImplementedError(  # pragma: no cover
+                        f"Unhandled append mode {mode}"
+                    )
             else:
                 LOGGER.warning('Appending disabled, dry run!')
+
+    def _append_with_insertions(self, ds):
+        append_dim = self._output_append_dim
+        for i in range(ds.dims[append_dim]):
+            dataslice = ds.isel({append_dim: slice(i, i+1)})
+            insert_index, update_mode = \
+                find_slice(self._output_store, dataslice[append_dim],
+                           dimension=append_dim)
+            if update_mode == "append":
+                append_slice(self._output_store, dataslice,
+                             dimension=append_dim)
+            elif update_mode == "insert":
+                # Currently only works for local filesystem stores
+                update_slice(self._output_store, insert_index,
+                             dataslice, update_mode, dimension=append_dim)
+            elif update_mode == "replace":
+                if self._output_append_mode is AppendMode.replace:
+                    update_slice(self._output_store, insert_index,
+                                 dataslice, update_mode, dimension=append_dim)
+                # If the append mode is not "replace", it must be "retain" --
+                # so we do nothing, which retains the existing slice and
+                # discards the new one.
+            else:
+                # The remaining possible value is "create" -- but we've
+                # already confirmed that we have an existing store,
+                # so "create" is a "can't happen".
+                raise NotImplementedError(  # pragma: no cover
+                    f"Unhandled update mode {update_mode}!")
+
+    def _check_append_allowed(self, ds: xr.Dataset) -> None:
+        output_ds = None
+        try:
+            output_ds = xr.open_zarr(self._output_store)
+        except GroupNotFoundError:
+            pass
+            # Output store doesn't exist, so we'll skip checks on it.
+            # (This should only happen in the case of a dry run, in cases
+            # where the output file would have been created in a normal
+            # run, so self._output_path_exists is true but the Zarr doesn't
+            # actually exist.)
+        if self._output_append_mode is not AppendMode.all:
+            if not self._is_append_dim_monotonic_increasing(output_ds):
+                raise ValueError(
+                    f"Existing {self._output_append_dim} values must "
+                    f"be increasing.")
+        if output_ds is not None and \
+                self._output_append_mode is AppendMode.no_overlap:
+            if output_ds[self._output_append_dim][-1] > \
+                    ds[self._output_append_dim][0]:
+                raise ValueError(
+                    f"Existing and appended {self._output_append_dim} "
+                    f"values may not overlap when using "
+                    f"{self._output_append_mode.name}.")
+        if self._output_append_mode is AppendMode.newer:
+            if not self._is_append_dim_monotonic_increasing(ds):
+                raise ValueError(
+                    f"Appended {self._output_append_dim} values must "
+                    f"be increasing when using "
+                    f"{self._output_append_mode.name}.")
+
+    def _is_append_dim_monotonic_increasing(self, ds: xr.Dataset) -> bool:
+        arr = ds[self._output_append_dim].data
+        return np.all(arr[1:] >= arr[:-1])
 
     def _remove_dataset(self):
         with log_duration(f'Removing dataset {self._output_path}'):
@@ -241,10 +345,12 @@ class DatasetWriter:
         source = None
         if self._input_paths:
             # Note, currently we only name root sources = our NetCDF files.
-            nc_paths = [path for path in self._input_paths if path.endswith('.nc')]
+            nc_paths = \
+                [path for path in self._input_paths if path.endswith('.nc')]
             if nc_paths:
                 source = dataset.attrs.get('source')
-                source = ((source + ',\n') if source else '') + ', '.join(nc_paths)
+                source = \
+                    ((source + ',\n') if source else '') + ', '.join(nc_paths)
         return source
 
     @classmethod