+++ tests/helpers.py
@@ -194,6 +194,20 @@ def new_test_dataset(w: int = 36,
     return dataset
 
 
+def new_append_test_datasets(dates1: List[str], dates2: List[str]) \
+        -> (xr.Dataset, xr.Dataset):
+    return xr.Dataset(
+        {"v": (["t", "x", "y"], np.zeros((len(dates1), 3, 3)))},
+        coords={"t": np.array(dates1, dtype="datetime64"),
+                "x": np.array([0, 1, 2]), "y": np.array([0, 1, 2])
+                }), \
+           xr.Dataset(
+        {"v": (["t", "x", "y"], np.ones((len(dates2), 3, 3)))},
+        coords={"t": np.array(dates2, dtype="datetime64"),
+                "x": np.array([0, 1, 2]), "y": np.array([0, 1, 2]),
+                })
+
+
 def delete_path(path, ignore_errors=False):
     if os.path.isdir(path):
         shutil.rmtree(path, ignore_errors=ignore_errors)